mlops-hm3-dvc/requirements.txt
dvc[s3]==3.47.0
pandas==2.2.0
scikit-learn==1.4.1.post1
pyyaml==6.0.1
pyarrow==15.0.0
dvclive==3.43.0
gto==1.7.0